AutoRCCar/requirements.txt
bleach==1.5.0
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
h5py==2.7.1
html5lib==0.9999999
imutils==0.4.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.0.8
Markdown==2.6.9
matplotlib==2.0.2
numpy==1.13.1+mkl
olefile==0.44
opencv-python==3.3.0.9
pandas==0.20.3
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.5
tornado==4.5.2
tqdm==4.15.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
